some issues on visual layer fixed images are now presented correctly
zooming is now available

+++ src/core/confighandler.cpp
@@ -186,7 +186,7 @@ void ConfigHandler::transactNewConfig(QMap<QString, QString> *temp_config_map)
         {
             this->config_map->insert(key,value);
             if(key == BACKGROUND_OPACITY_SWITCH)
-                emit this->backgroundOpacityChanged(value.toDouble());  //update the opacity in the QmlInterface
+                emit this->backgroundOpacityChanged(value.toFloat());  //update the opacity in the QmlInterface
             if(key == BACKGROUND_COLOR_SWITCH)
                 emit this->backgroundColorChanged(value);               //update the color in the QmlInterface
             if(key == FILE_FORMAT_BLACKLIST_SWITCH)
